+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [v0.0.2] - 2026-02-11
+
+- Fix PostHog telemetry
+- Add Pipelex install skills when installing methods
+
 ## [v0.0.1] - 2026-02-11
 
 - Initial commit!